% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{deploy_filesystem}
\alias{deploy_filesystem}
\title{Automatically deploy the filysystem}
\usage{
deploy_filesystem()
}
\description{
This function will be called once when the package is loaded.
It will mount a filesystem to a temporary directory. The Travel
package relies on the filesystem to work. It is a no-op if the
filesystem has been mounted.
}
\examples{
deploy_filesystem()
}
