% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{deploy_filesystem}
\alias{deploy_filesystem}
\alias{get_mountpoint}
\alias{set_mountpoint}
\alias{run_filesystem}
\alias{stop_filesystem}
\alias{is_filesystem_running}
\title{Filesystem management}
\usage{
deploy_filesystem()

get_mountpoint()

set_mountpoint(path)

run_filesystem()

stop_filesystem()

is_filesystem_running()
}
\value{
Character vector or no return value
}
\description{
These functions manage the filesystem which Travel is used to
create virtual pointers. You should not call these functions
manually for Travel will automatically mount the filesystem
if it needs one.
}
\examples{
 
## Automatically determine the mountpoint
## and deploy the filesystem
deploy_filesystem()

## Get mountpoint for the current filesystem
get_mountpoint()
}
