% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer_APIs.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Find path of the Travel header file}
\usage{
pkgconfig(x = c("PKG_LIBS", "PKG_CPPFLAGS"))
}
\arguments{
\item{x}{Character, "PKG_LIBS" or "PKG_CPPFLAGS"}
}
\value{
path to the header or compiler flags
}
\description{
This function will return the path of the Travel header or
the flags that are used to compile C++ code
for the developers who want to use C++ level implementation of the
\code{Travel} package
}
\examples{
Travel:::pkgconfig("PKG_LIBS")
Travel:::pkgconfig("PKG_CPPFLAGS")
}
