% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_function.R
\name{travel_function}
\alias{travel_function}
\title{A Travel Function}
\usage{
travel_function(love = TRUE)
}
\arguments{
\item{love}{Do you love to travel? Defaults to TRUE.}
}
\description{
This function allows you to express your love of Travel
}
\examples{
travel_function()
}
\keyword{travel}
