% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altptr.R
\name{wrap_altrep}
\alias{wrap_altrep}
\title{Create a new ALTREP object from an R ALTREP}
\usage{
wrap_altrep(x)
}
\arguments{
\item{x}{an ALTREP object}
}
\value{
A new ALTREP object
}
\description{
Create a new ALTREP object from an R ALTREP, the new ALTREP object
owns a data pointer. However, it is a dummy pointer in the sense
that the data is not in the memory. The pointer can be operated as
a normal pointer but the request of accessing the data will be send to the
wrapped ALTREP object.
}
\examples{
x <- 1:10
y <- wrap_altrep(x)
}
